refactor(linter): split theme rules domain (Task 2.4)
- split 313 lines of theme.rs into logical submodules
- extract unused, hardcode, builder modules
- establish theme orchestration module
