+++ openspec/changes/v0-8-4-rust-code-refactoring/tasks.md
@@ -92,7 +92,8 @@ Tasks Grouped by ## = Adhere unconditionally to the branching standard defined i
 - [x] 2.3 `rules/i18n.rs`（391行）の分割
   - 責務ごとにサブモジュール化
 
-- [ ] 2.4 `rules/theme.rs`（313行）の分割
+- [x] 2.4 `rules/theme.rs`（313行）の分割
+  - 責務ごとにサブモジュール化
 
 - [ ] 2.5 `rules/markdown.rs`（204行）のボーダーライン確認・必要に応じて分割
